// samples/ContosoIntranet/ContosoIntranet/src/Models/ContosoNewsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */

export interface ContosoNews {
  ID?: number;
  Title: string;
  OData__ColorTag?: string;
  ComplianceAssetId?: string;
  Summary?: string;
  ImageUrl?: string;
  "Category#Id"?: number;
  Category?: Record<string, unknown>;
  PublishDate?: string;
  Modified?: string;
  Created?: string;
  "Author#Claims"?: string;
  Author?: Record<string, unknown>;
  "Editor#Claims"?: string;
  Editor?: Record<string, unknown>;
  "{Identifier}"?: string;
  "{IsFolder}"?: boolean;
  "{Thumbnail}"?: Record<string, unknown>;
  "{Link}"?: string;
  "{Name}"?: string;
  "{FilenameWithExtension}"?: string;
  "{Path}"?: string;
  "{FullPath}"?: string;
  "{ModerationStatus}"?: string;
  "{ModerationComment}"?: string;
  "{ContentType}#Id"?: string;
  "{ContentType}"?: Record<string, unknown>;
  "{HasAttachments}"?: boolean;
  "{Attachments}@odata.type"?: string;
  "{Attachments}"?: unknown;
  "{VersionNumber}"?: string;
  "{TriggerWindowStartToken}"?: string;
  "{TriggerWindowEndToken}"?: string;
}